requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
mysqlclient==2.1.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
PyMySQL==1.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
Flask-Login==0.6.3


